+++ src/Data/Aeson/JSONPath/Parser.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE DeriveLift #-}
 {-# OPTIONS_GHC -Wno-unused-do-bind #-}
 module Data.Aeson.JSONPath.Parser
   ( JSPQuery (..)
@@ -17,25 +18,26 @@ import Data.Functor                  (($>))
 import Data.Text                     (Text)
 import Data.Char                     (ord)
 import Text.ParserCombinators.Parsec ((<|>))
+import Language.Haskell.TH.Syntax    (Lift (..))
 
 import Prelude
 
 data JSPQuery
   = JSPRoot [JSPSegment]
-  deriving (Eq, Show)
+  deriving (Eq, Show, Lift)
 
 -- https://www.rfc-editor.org/rfc/rfc9535#name-segments-2
 data JSPSegment
   = JSPChildSeg JSPChildSegment
   | JSPDescSeg JSPDescSegment
-  deriving (Eq, Show)
+  deriving (Eq, Show, Lift)
 
 -- https://www.rfc-editor.org/rfc/rfc9535#name-child-segment
 data JSPChildSegment
   = JSPChildBracketed [JSPSelector]
   | JSPChildMemberNameSH JSPNameSelector
   | JSPChildWildSeg JSPWildcardT
-  deriving (Eq, Show)
+  deriving (Eq, Show, Lift)
 
 
 -- https://www.rfc-editor.org/rfc/rfc9535#name-descendant-segment
@@ -43,7 +45,7 @@ data JSPDescSegment
   = JSPDescBracketed [JSPSelector]
   | JSPDescMemberNameSH JSPNameSelector
   | JSPDescWildSeg JSPWildcardT
-  deriving (Eq, Show)
+  deriving (Eq, Show, Lift)
 
 -- https://www.rfc-editor.org/rfc/rfc9535#name-selectors-2
 data JSPSelector
@@ -51,10 +53,10 @@ data JSPSelector
   | JSPIndexSel JSPIndexSelector
   | JSPSliceSel JSPSliceSelector
   | JSPWildSel JSPWildcardT
-  deriving (Eq, Show)
+  deriving (Eq, Show, Lift)
 
 data JSPWildcardT = JSPWildcard
-  deriving (Eq, Show)
+  deriving (Eq, Show, Lift)
 
 type JSPNameSelector = Text