+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix parsing bug with Wildcard Selector
 - Implement Descendant Segment
 - Fix allowed characters in the `member-name-shorthand`
+- Add `QuasiQuoter` for compile-time syntax checking
 
 ## 0.1.0.0